fix: address mchmarny review round 2 (readiness, lint cleanup)
Address the May 11 PR #469 follow-up review.

Readiness-bind race (the only fully-unaddressed item from round 2):

  pkg/server/server.go Start() previously kicked off ListenAndServe in
  a goroutine and immediately flipped readiness on the main goroutine,
  opening a window where /ready returned 200 while the kernel was still
  binding the socket. Probes during that window would 200 then get
  connection-refused on the next request.

  Bind the listener synchronously (net.ListenConfig.Listen — context-
  aware, satisfies noctx) BEFORE setReady(true), then start
  http.Server.Serve(ln) in a goroutine. If the bind fails the server
  never advertises ready. New regression test
  TestReadyImpliesListenerBound asserts that a dial against the port
  succeeds the moment ready flips to true.

golangci-lint failures cleared (the CI gate this PR currently fails on):

- exhaustive (3 sites): pkg/errors/exitcode.go, pkg/server/errors.go —
  ErrCodePolicyFetch is now covered in the exit-code, HTTP-status, and
  retryable switches. Treated as upstream-unavailable (ExitUnavailable
  / 503 / retryable=true) since policy fetch hits an external URI.

- gosec G115 (int→uint16): pkg/server/config.go ResolvePort now
  returns uint16 directly, matching net's port representation. Single
  cast still needed in parseConfig because config.Port is int for
  back-compat with existing tests that set cfg.Port = port; gated with
  a nolint comment since ResolvePort guarantees the range.
  pkg/api/server.go drops the int→uint16 cast at the registration site.

- gosec G704 (SSRF taint): pkg/openshell/fetcher.go Fetch annotated
  with nolint pointing at the three layers already mitigating the
  finding (URL pre-flight, post-resolution dial guard, redirect
  check) and the regression test that exercises the dial path. Also
  unblocks the CodeQL job which was firing on the same taint analysis.

- noctx: internal/testdns/testdns.go Start switches from net.ListenPacket
  to (*net.ListenConfig).ListenPacket with context.Background() per
  noctx lint requirement.

Layer 1 strict-mode policy fields remain a documented limitation in
Guard's doc comment (per the reviewer's "downgraded to documented
limitation rather than fixed" note). Tracked in a follow-up issue.

+++ pkg/api/server.go
@@ -211,7 +211,7 @@ func setupDiscovery() ([]server.Option, http.HandlerFunc, error) {
 		Protocol:     discovery.ProtocolMCP,
 		Namespace:    ns,
 		ServiceName:  name,
-		Port:         uint16(server.ResolvePort(defaults.ServerDefaultPort)),
+		Port:         server.ResolvePort(defaults.ServerDefaultPort),
 		Capabilities: []string{"recipe", "bundle", "validate", "query"},
 		Version:      version,
 	}
